fix validation

+++ ylab/src/main/java/ru/dariayo/controllers/CarController.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import jakarta.validation.Valid;
 import ru.dariayo.dto.CarDTO;
 import ru.dariayo.mapper.CarMapper;
 import ru.dariayo.model.Car;
@@ -34,24 +35,30 @@ public class CarController {
     }
 
     @PostMapping
-    public ResponseEntity<CarDTO> addCar(@RequestBody CarDTO carDTO) {
+    public ResponseEntity<CarDTO> addCar(@Valid @RequestBody CarDTO carDTO) {
         Car car = carMapper.carDTOToCar(carDTO);
         carRepository.addCar(car);
         return ResponseEntity.status(HttpStatus.CREATED).body(carDTO);
     }
 
     @PutMapping("/update/{mark}/{model}")
-    public ResponseEntity<CarDTO> updateCar(
+    public ResponseEntity<?> updateCar(
             @PathVariable String mark,
             @PathVariable String model,
-            @RequestBody CarDTO updatedCarDTO) {
+            @Valid @RequestBody CarDTO updatedCarDTO) {
+        if (mark == null || model == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Request body is missing");
+        }
         Car updatedCar = carMapper.carDTOToCar(updatedCarDTO);
         carRepository.updateCar(mark, model, updatedCar);
         return ResponseEntity.ok(updatedCarDTO);
     }
 
     @DeleteMapping("/remove/{mark}/{model}")
-    public ResponseEntity<Void> removeCar(@PathVariable String mark, @PathVariable String model) {
+    public ResponseEntity<?> removeCar(@PathVariable String mark, @PathVariable String model) {
+        if (mark == null || model == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Request body is missing");
+        }
         boolean result = carRepository.removeCar(mark, model);
         if (result) {
             return ResponseEntity.noContent().build();
@@ -61,7 +68,10 @@ public class CarController {
     }
 
     @GetMapping("/search")
-    public ResponseEntity<List<CarDTO>> searchCar(@RequestParam String param, @RequestParam String value) {
+    public ResponseEntity<?> searchCar(@RequestParam String param, @RequestParam String value) {
+        if (param == null || value == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Request body is missing");
+        }
         List<CarDTO> carDtos = carRepository.searchCar(param, value).stream()
                 .map(carMapper::carToCarDTO)
                 .collect(Collectors.toList());
